# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "GSES BTC application",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "gses2.app:8080",
    "basePath": "/api",
    "paths": {
        "/rate": {
            "get": {
                "description": "Запит має повертати поточний курс USD до UAH використовуючи будь-який third party сервіс з публічним АРІ",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rate"
                ],
                "summary": "Отримати поточний курс USD до UAH",
                "operationId": "rate",
                "responses": {
                    "200": {
                        "description": "Повертається актуальний курс USD до UAH",
                        "schema": {
                            "type": "number"
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/subscribe": {
            "post": {
                "description": "Запит має перевірити, чи немає данної електронної адреси в поточній базі даних і, в разі її відсутності, записувати її.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Підписати емейл на отримання поточного курсу",
                "operationId": "subscribe",
                "parameters": [
                    {
                        "type": "string",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "E-mail додано"
                    },
                    "409": {
                        "description": "Повертати, якщо e-mail вже є в базі даних"
                    }
                }
            }
        }
    }
}